+++ src/main/java/com/github/lemniscate/spring/crud/annotation/ApiController.java
@@ -8,6 +8,6 @@ import java.lang.annotation.*;
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
 @Component
-public @interface ApiResourceController {
+public @interface ApiController {
 
 }