+++ src/main/java/com/github/lemniscate/spring/crud/annotation/EnableApiResources.java
@@ -1,7 +1,9 @@
 package com.github.lemniscate.spring.crud.annotation;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.lemniscate.spring.crud.svc.ApiResourceServices;
 import com.github.lemniscate.spring.crud.util.ApiResourceRegistry;
+import com.github.lemniscate.spring.crud.view.JsonViewResolver;
 import com.github.lemniscate.spring.crud.web.assembler.ApiResourceAssemblers;
 import com.github.lemniscate.spring.crud.web.assembler.ApiResourceEntityLinks;
 import com.github.lemniscate.spring.crud.web.assembler.ApiResourceLinkBuilderFactory;
@@ -11,8 +13,8 @@ import org.springframework.beans.factory.support.AbstractBeanDefinition;
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
 import org.springframework.beans.factory.support.BeanDefinitionRegistry;
 import org.springframework.beans.factory.support.RootBeanDefinition;
-import org.springframework.context.annotation.Import;
-import org.springframework.context.annotation.ImportBeanDefinitionRegistrar;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.*;
 import org.springframework.core.annotation.AnnotationAttributes;
 import org.springframework.core.type.AnnotationMetadata;
 import org.springframework.core.type.StandardAnnotationMetadata;
@@ -22,7 +24,7 @@ import java.lang.annotation.*;
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-@Import(EnableApiResources.ApiResourceRegistrar.class)
+@Import({EnableApiResources.ApiResourceRegistrar.class, EnableApiResources.ApiResourceConfiguration.class})
 public @interface EnableApiResources {
     Class<?> value() default EnableApiResources.class;
 
@@ -60,5 +62,16 @@ public @interface EnableApiResources {
             registry.registerBeanDefinition("apiResourcesLinkBuilderFactory", new RootBeanDefinition(ApiResourceLinkBuilderFactory.class));
             registry.registerBeanDefinition("apiResourcesEntityLinks", new RootBeanDefinition(ApiResourceEntityLinks.class));
         }
+    }
+
+    @Configuration
+    public class ApiResourceConfiguration{
+
+        @Bean
+        @ConditionalOnMissingBean(JsonViewResolver.class)
+        public JsonViewResolver jsonViewResolver(){
+            return new JsonViewResolver();
+        }
+
     }
 }