+++ src/main/java/com/github/lemniscate/spring/crud/web/ApiResourceController.java
@@ -4,7 +4,9 @@ import com.github.lemniscate.spring.crud.mapping.ApiResourceMapping;
 import com.github.lemniscate.spring.crud.svc.ApiResourceService;
 import com.github.lemniscate.spring.crud.util.ApiResourceSupport;
 import com.github.lemniscate.spring.crud.util.ApiResourceUtil;
+import com.github.lemniscate.spring.crud.view.JsonViewResolver;
 import com.github.lemniscate.spring.crud.web.assembler.ApiResourceAssembler;
+import com.github.lemniscate.spring.jsonviews.client.BaseView;
 import lombok.Getter;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -12,6 +14,7 @@ import org.springframework.hateoas.Identifiable;
 import org.springframework.hateoas.Resource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,8 +25,12 @@ import java.util.Map;
 /**
 * @Author dave 8/8/14 9:20 PM
 */
+@Transactional
 public class ApiResourceController<ID extends Serializable, E extends Identifiable<ID>, CB, RB extends Identifiable<ID>, UB>
         extends ApiResourceSupport {
+
+    @Inject
+    protected JsonViewResolver jsonViewResolver;
 
     @Inject
     protected ApiResourceService<ID, E, CB, RB, UB> service;
@@ -50,13 +57,13 @@ public class ApiResourceController<ID extends Serializable, E extends Identifiab
     @RequestMapping(value="", method= RequestMethod.GET)
     public ResponseEntity<Page<Resource<RB>>> getAll(@RequestParam MultiValueMap<String, String> params, Pageable p){
         Page<RB> entities = service.findForRead(p);
-        return assemblers.respond(mapping.readBeanClass(), entities, p, HttpStatus.OK);
+        return assemblers.respondWithView(mapping.readBeanClass(), view(ControllerMethod.GET_ALL), entities, p, HttpStatus.OK);
     }
 
     @RequestMapping(value="/{id}", method=RequestMethod.GET)
     public ResponseEntity<Resource<RB>> getOne(@PathVariable ID id){
         RB entity = service.read(conversionService.convert(id, mapping.idClass()));
-        return assemblers.respond(entity, entity == null ? HttpStatus.NOT_FOUND : HttpStatus.OK);
+        return assemblers.respondWithView(entity, view(ControllerMethod.GET_ONE), entity == null ? HttpStatus.NOT_FOUND : HttpStatus.OK);
     }
 
     @RequestMapping(value="/{id}", method=RequestMethod.DELETE)
@@ -68,19 +75,23 @@ public class ApiResourceController<ID extends Serializable, E extends Identifiab
     @RequestMapping(value="", method=RequestMethod.POST)
     public ResponseEntity<Resource<RB>> postOne(@RequestBody CB bean){
         RB entity = service.createForRead(bean);
-        return assemblers.respond(entity, HttpStatus.CREATED);
+        return assemblers.respondWithView(entity, view(ControllerMethod.POST), HttpStatus.CREATED);
     }
 
     @RequestMapping(value="/{id}", method=RequestMethod.PUT)
     public ResponseEntity<Resource<RB>> putOne(@PathVariable ID id, @RequestBody UB bean){
         RB entity = service.updateForRead( id, bean );
-        return assemblers.respond(entity, entity == null ? HttpStatus.NOT_FOUND : HttpStatus.OK);
+        return assemblers.respondWithView(entity, view(ControllerMethod.PUT), entity == null ? HttpStatus.NOT_FOUND : HttpStatus.OK);
     }
 
     @RequestMapping(value="/searches", method=RequestMethod.POST)
     public ResponseEntity<Page<Resource<RB>>> search(@RequestBody Map<String, Object> search, Pageable pageable){
         Page<RB> entities = service.searchForRead(search, pageable);
-        return assemblers.respond( mapping.readBeanClass(), entities, pageable, HttpStatus.OK);
+        return assemblers.respondWithView( mapping.readBeanClass(), view(ControllerMethod.SEARCH), entities, pageable, HttpStatus.OK);
+    }
+
+    private Class<? extends BaseView> view(ControllerMethod method){
+        return jsonViewResolver.resolve(method, mapping.readBeanClass());
     }
 
 }