+++ src/main/java/com/github/lemniscate/spring/crud/web/ControllerMethod.java
@@ -1,0 +1,13 @@
+package com.github.lemniscate.spring.crud.web;
+
+/**
+ * @Author dave 2/1/15 1:09 PM
+ */
+public enum ControllerMethod {
+    GET_ALL,
+    GET_ONE,
+    DELETE,
+    POST,
+    PUT,
+    SEARCH
+}